silence warnings about float/integer conversion

+++ src/smallmap_gui.cpp
@@ -845,10 +845,13 @@ public:
 						if (!show_towns) {
 							typedef std::multimap<uint, byte, std::greater<uint> > SizeMap;
 							SizeMap sizes;
-							sizes.insert(std::make_pair(sqrt((float)ls.usage), _colour_gradient[COLOUR_GREY][1]));
-							sizes.insert(std::make_pair(sqrt((float)ls.capacity), _colour_gradient[COLOUR_WHITE][7]));
-							sizes.insert(std::make_pair(sqrt((float)flow.planned),  _colour_gradient[COLOUR_RED][5]));
-							sizes.insert(std::make_pair(sqrt((float)flow.sent), _colour_gradient[COLOUR_YELLOW][5]));
+							/* these floats only serve to calculate the size of the coloured boxes for capacity, usage, planned, sent
+							 * they are not reused anywhere, so it's network safe.
+							 */
+							sizes.insert(std::make_pair((uint)sqrt((float)ls.usage), _colour_gradient[COLOUR_GREY][1]));
+							sizes.insert(std::make_pair((uint)sqrt((float)ls.capacity), _colour_gradient[COLOUR_WHITE][7]));
+							sizes.insert(std::make_pair((uint)sqrt((float)flow.planned),  _colour_gradient[COLOUR_RED][5]));
+							sizes.insert(std::make_pair((uint)sqrt((float)flow.sent), _colour_gradient[COLOUR_YELLOW][5]));
 
 							ptm.x = (pta.x + ptb.x) / 2;
 							ptm.y = (pta.y + ptb.y) / 2;
